+++ DCManagement/functions/shares/Get-DCShare.ps1
@@ -1,5 +1,21 @@
 ﻿function Get-DCShare
 {
+<#
+	.SYNOPSIS
+		Returns the list of registered shares.
+	
+	.DESCRIPTION
+		Returns the list of registered shares.
+	
+	.PARAMETER Name
+		Filter the returned share definitions by name.
+		Defaults to '*'
+	
+	.EXAMPLE
+		PS C:\> Get-DCShare
+	
+		Returns the list of registered shares.
+#>
 	[CmdletBinding()]
 	Param (
 		[string]