+++ DCManagement/functions/shares/Register-DCShare.ps1
@@ -1,5 +1,5 @@
 ﻿function Register-DCShare {
-	<#
+<#
 	.SYNOPSIS
 		Registers an SMB share that should exist on DCs.
 	
@@ -14,14 +14,21 @@
 		The path the share points to.
 		Supports string resolution.
 	
+	.PARAMETER Description
+		The description of the share.
+		Supports string resolution.
+	
 	.PARAMETER FullAccess
 		The principals to grant full access to.
+		Supports string resolution.
 	
 	.PARAMETER WriteAccess
 		The principals to grant write access to.
+		Supports string resolution.
 	
 	.PARAMETER ReadAccess
 		The principals to grant read access to.
+		Supports string resolution.
 	
 	.PARAMETER AccessMode
 		How share access rules are processed.
@@ -36,11 +43,16 @@
 		- FSMO: Only DCs that have any FSMO role
 		- PDC:  Only the PDCEmulator
 	
+	.PARAMETER ContextName
+		The name of the context defining the setting.
+		This allows determining the configuration set that provided this setting.
+		Used by the ADMF, available to any other configuration management solution.
+	
 	.EXAMPLE
 		PS C:\> Get-Content .\shares.json | ConvertFrom-Json | Write-Output | Register-DCShare
-
+		
 		Reads all share definitions from json and imports the definitions.
-	#>
+#>
 	[CmdletBinding()]
 	Param (
 		[Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
@@ -50,7 +62,12 @@
 		[Parameter(Mandatory = $true, ValueFromPipelineByPropertyName = $true)]
 		[string]
 		$Path,
-
+		
+		[Parameter(ValueFromPipelineByPropertyName = $true)]
+		[AllowEmptyCollection()]
+		[string]
+		$Description,
+		
 		[Parameter(ValueFromPipelineByPropertyName = $true)]
 		[AllowEmptyCollection()]
 		[string[]]
@@ -74,19 +91,24 @@
 		[Parameter(ValueFromPipelineByPropertyName = $true)]
 		[ValidateSet('All', 'FSMO', 'PDC')]
 		[string]
-		$ServerRole = 'All'
+		$ServerRole = 'All',
+		
+		[string]
+		$ContextName = '<Undefined>'
 	)
 	
 	process {
 		$script:shares[$Name] = [PSCustomObject]@{
 			PSTypeName  = 'DCManagement.Share'	
 			Name        = $Name
-			Path        = $Path
+			Path	    = $Path
+			Description = $Description
 			FullAccess  = $FullAccess
 			WriteAccess = $WriteAccess
 			ReadAccess  = $ReadAccess
 			AccessMode  = $AccessMode
 			ServerRole  = $ServerRole
+			ContextName = $ContextName
 		}
 	}
 }