+++ DCManagement/changelog.md
@@ -1,5 +1,10 @@
 ﻿# Changelog
 
+## ???
+
+- New Component: Shares
+- New Command: Clear-DCConfiguration
+
 ## 1.0.5 (2020-01-27)
 
 - Metadata Update, no functional changes